feat: ensure ChangeProfilePicture use case doesn't call DeleteFile when file is not valid and SaveUserPicture throws

+++ src/domain/useCases/changeProfilePicture.ts
@@ -25,7 +25,8 @@ export class ChangeProfilePictureUseCase implements ChangeProfilePicture {
     try {
       await this.userProfileRepository.savePicture(userProfile)
     } catch (error) {
-      await this.fileStorage.delete({ key })
+      if (file) await this.fileStorage.delete({ key })
+      throw new Error()
     }
     return userProfile
   }
